fix(kotlin): iconSize expects Int, not Float (CI build failure)
CI run https://github.com/hoshiyomiX/OTAku/actions/runs/28293742994
failed on all 3 APK build jobs (ARMv7, ARMv8, Universal) at step
"Build release APK" with:

  e: MainActivity.kt:1396:32 Type mismatch: inferred type is Float
     but Int was expected
  Execution failed for task ':app:compileArm8ReleaseKotlin'.

Root cause: MaterialButton.setIconSize() expects Int (pixels), but the
previous commit passed `dpToPx(18).toFloat()` (a Float).

The standalone kotlinc syntax check in the sandbox missed this because
kotlinc without the Material library on its classpath only checks
parser errors, not type errors against external APIs. The full Gradle
build with the Material dependency correctly caught the type mismatch.

Fix: Remove the .toFloat() — dpToPx() already returns Int.

File: android/app/src/main/java/com/hoshiyomi/otaku/MainActivity.kt
Line: 1396 → iconSize = dpToPx(18)  (was: dpToPx(18).toFloat())

Verification:
- Kotlin syntax check: PASS (0 syntax errors)
- Type signature verified: dpToPx(dp: Int): Int matches
  MaterialButton.setIconSize(size: Int)

+++ android/app/src/main/java/com/hoshiyomi/otaku/MainActivity.kt
@@ -1393,7 +1393,11 @@ class MainActivity : AppCompatActivity() {
                     icon = androidx.core.content.ContextCompat.getDrawable(
                         this@MainActivity, R.drawable.ic_close
                     )
-                    iconSize = dpToPx(18).toFloat()
+                    // MaterialButton.setIconSize() expects Int (pixels), not Float.
+                    // dpToPx() already returns Int — no .toFloat() needed.
+                    // (Previous .toFloat() caused CI build failure:
+                    // "Type mismatch: inferred type is Float but Int was expected")
+                    iconSize = dpToPx(18)
                     text = null  // icon-only button
                     insetTop = 0
                     insetBottom = 0
